moved WireDriveFileUITracker to separate file
