+++ WireMessaging/Sources/WireMessagingDomain/WireDrive/Model/WireDriveLocalAsset.swift
@@ -125,88 +125,3 @@ package extension WireDriveLocalAsset.DownloadState {
         }
     }
 }
-
-//TODO: move this type to its own file
-
-/// Intended for tracking the UI state of Drive files in different parts of the app.
-/// Consumes state changes from `WireDriveLocalAsset.DownloadState` and provides observable state changes specifically for the UI that represents the file.
-/// Also notifies if and when a file should be opened automatically after download.
-@MainActor
-public final class WireDriveFileUITracker: Sendable, ObservableObject {
-    public enum State: Sendable {
-        /// The file hasn't been downloaded yet.
-        case notDownloaded
-        
-        /// The file is currently being downloaded.
-        case downloading(progress: Double)
-        
-        /// The file has been downloaded.
-        /// `downloaded(showReadyToOpen: true)` indicates that the file should be shown in a specific "ready to open" state in the UI.
-        /// Will automatically change to `downloaded(showReadyToOpen: false)` state after a few seconds.
-        case downloaded(showReadyToOpen: Bool)
-        
-        /// The download has failed.
-        case failed(error: any Error)
-    }
-    
-    private(set) public var state: State {
-        didSet {
-            switch (oldValue, state) {
-            case (.downloading, .downloaded):
-                // state is changing from downloading to downloaded:
-                
-                // if the download takes longer than this amount of time, the file is considered "big"
-                let downloadTimeForBigFiles: TimeInterval = 0.3
-                
-                let fileIsBig = downloadingTimerStart.flatMap { Date().timeIntervalSince($0) > downloadTimeForBigFiles } ?? true
-                
-                self.state = .downloaded(showReadyToOpen: fileIsBig)
-                
-                if fileIsBig {
-                    Task {
-                        try? await Task.sleep(for: .seconds(3))
-                        self.state = .downloaded(showReadyToOpen: false)
-                    }
-                } else {
-                    fileShouldOpen?()
-                }
-            default:
-                break
-            }
-            
-            if case .downloading = state {
-                if downloadingTimerStart == nil {
-                    downloadingTimerStart = Date()
-                }
-            } else {
-                downloadingTimerStart = nil
-            }
-        }
-    }
-    
-    /// This closure will be called when a file should be automatically opened after the download.
-    public var fileShouldOpen: (() -> ())?
-    
-    private var downloadingTimerStart: Date? = nil
-    
-    public init(downloadState: WireDriveLocalAsset.DownloadState) {
-        state = Self.stateFromDownloadState(downloadState)
-    }
-    
-    public func handleDownloadState(_ downloadState: WireDriveLocalAsset.DownloadState) {
-        state = Self.stateFromDownloadState(downloadState)
-    }
-    
-    private static func stateFromDownloadState(_ downloadState: WireDriveLocalAsset.DownloadState) -> State {
-        switch downloadState {
-        case .pending:
-            .notDownloaded
-        case let .downloading(progress):
-            .downloading(progress: progress)
-        case .downloaded:
-            .downloaded(showReadyToOpen: false)
-        case let .failed(error):
-            .failed(error: error)
-        }
-    }
-}
